requirements.txt
discord.py==1.3.3
requests==2.23.0
pymongo[tls, srv]==3.10.1
google-api-python-client==1.9.1
Pillow==7.1.2